\name{rbndb}
\docType{package}
\title{rbndb: Download Data from BNDB Ecuador}
\keyword{package}

 DESCRIPTION{
R package to download occurrence data from BNDB Ecuador.
}

\section{Functions}{

download_bndb

}

\section{Arguments}{

scientific_name: Species name
max_pages: Max pages
delay: Delay seconds
polygon: Polygon path
crs: Coordinate system
output: csv or shp
map: TRUE or FALSE
out_file: Output file

}
