% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bndb.R
\name{rbndb}
\alias{rbndb}
\title{rbndb: Download Data from BNDB Ecuador}
\docType{package}
\usage{
library(rbndb)
}
DESCRIPTION{
R package to download occurrence data from BNDB Ecuador via web scraping.
}
\section{Functions}{
\code{download_bndb}: Downloads occurrence data from BNDB.
}
\section{Examples}{
library(rbndb)
occ <- download_bndb("Cedrela odorata")
}
\keyword{package}
